[COMMON] Entity 설계

+++ src/main/java/kr/co/iabacus/sales/web/common/converter/RatioConverter.java
@@ -1,20 +1,22 @@
 package kr.co.iabacus.sales.web.common.converter;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 import kr.co.iabacus.sales.web.common.Ratio;
 
 @Converter(autoApply = true)
-public class RatioConverter implements AttributeConverter<Ratio, Double> {
+public class RatioConverter implements AttributeConverter<Ratio, BigDecimal> {
 
     @Override
-    public Double convertToDatabaseColumn(Ratio ratio) {
+    public BigDecimal convertToDatabaseColumn(Ratio ratio) {
         return ratio.getRate();
     }
 
     @Override
-    public Ratio convertToEntityAttribute(Double rate) {
+    public Ratio convertToEntityAttribute(BigDecimal rate) {
         return Ratio.valueOf(rate);
     }
